+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/composer/MessageComposer.kt
@@ -1,6 +1,8 @@
 package io.getstream.chat.android.compose.ui.messages.composer
 
 import android.widget.Toast
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
@@ -8,6 +10,7 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
 import androidx.compose.material.Surface
@@ -16,15 +19,19 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Alignment.Companion.Bottom
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.getstream.sdk.chat.utils.MediaStringUtil
 import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.client.models.Message
+import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.common.state.Edit
 import io.getstream.chat.android.common.state.ValidationError
 import io.getstream.chat.android.compose.R
@@ -32,6 +39,7 @@ import io.getstream.chat.android.compose.state.messages.composer.MessageComposer
 import io.getstream.chat.android.compose.ui.messages.composer.components.DefaultComposerIntegrations
 import io.getstream.chat.android.compose.ui.messages.composer.components.MessageInput
 import io.getstream.chat.android.compose.ui.messages.composer.components.MessageInputOptions
+import io.getstream.chat.android.compose.ui.messages.suggestions.mentions.MentionSuggestionList
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 import io.getstream.chat.android.compose.viewmodel.messages.MessageComposerViewModel
 
@@ -48,6 +56,8 @@ import io.getstream.chat.android.compose.viewmodel.messages.MessageComposerViewM
  * @param onValueChange Handler when the input field value changes.
  * @param onAttachmentRemoved Handler when the user taps on the cancel/delete attachment action.
  * @param onCancelAction Handler for the cancel button on Message actions, such as Edit and Reply.
+ * @param onMentionClick @param onMentionClick Handler when the user taps on a mention suggestion item.
+ * @param mentionPopupContent Customizable composable function that represents the mention suggestions popup.
  * @param integrations A view that represents custom integrations. By default, we provide
  * [DefaultComposerIntegrations], which show Attachments & Giphy, but users can override this with
  * their own integrations, which they need to hook up to their own data providers and UI.
@@ -63,6 +73,13 @@ public fun MessageComposer(
     onValueChange: (String) -> Unit = { viewModel.setMessageInput(it) },
     onAttachmentRemoved: (Attachment) -> Unit = { viewModel.removeSelectedAttachment(it) },
     onCancelAction: () -> Unit = { viewModel.dismissMessageActions() },
+    onMentionClick: (User) -> Unit = { viewModel.selectMention(it) },
+    mentionPopupContent: @Composable (List<User>) -> Unit = {
+        DefaultMentionPopupContent(
+            mentionSuggestions = it,
+            onMentionClick = onMentionClick
+        )
+    },
     integrations: @Composable RowScope.() -> Unit = {
         DefaultComposerIntegrations(onAttachmentsClick)
     },
@@ -84,6 +101,8 @@ public fun MessageComposer(
     val selectedAttachments by viewModel.selectedAttachments.collectAsState()
     val activeAction by viewModel.lastActiveAction.collectAsState(null)
     val validationErrors by viewModel.validationErrors.collectAsState()
+    val mentionSuggestions by viewModel.mentionSuggestions.collectAsState()
+    val cooldownTimer by viewModel.cooldownTimer.collectAsState()
 
     MessageComposer(
         modifier = modifier,
@@ -92,13 +111,17 @@ public fun MessageComposer(
 
             onSendMessage(messageWithData)
         },
+        onMentionClick = onMentionClick,
+        mentionPopupContent = mentionPopupContent,
         integrations = integrations,
         input = input,
         messageComposerState = MessageComposerState(
             inputValue = value,
             attachments = selectedAttachments,
             action = activeAction,
-            validationErrors = validationErrors
+            validationErrors = validationErrors,
+            mentionSuggestions = mentionSuggestions,
+            cooldownTimer = cooldownTimer,
         ),
         shouldShowIntegrations = true,
         onCancelAction = onCancelAction
@@ -112,7 +135,9 @@ public fun MessageComposer(
  * @param messageComposerState The state of the message input.
  * @param onSendMessage Handler when the user taps on the send message button.
  * @param onCancelAction Handler when the user cancels the active action (Reply or Edit).
+ * @param onMentionClick Handler when the user taps on a mention suggestion item.
  * @param modifier Modifier for styling.
+ * @param mentionPopupContent Customizable composable function that represents the mention suggestions popup.
  * @param shouldShowIntegrations If we should show or hide integrations.
  * @param integrations Composable that represents integrations for the composer, such as Attachments.
  * @param input Composable that represents the input field in the composer.
@@ -122,14 +147,21 @@ public fun MessageComposer(
     messageComposerState: MessageComposerState,
     onSendMessage: (String, List<Attachment>) -> Unit,
     onCancelAction: () -> Unit,
+    onMentionClick: (User) -> Unit,
     modifier: Modifier = Modifier,
+    mentionPopupContent: @Composable (List<User>) -> Unit = {
+        DefaultMentionPopupContent(
+            mentionSuggestions = it,
+            onMentionClick = onMentionClick
+        )
+    },
     shouldShowIntegrations: Boolean = true,
     integrations: @Composable RowScope.() -> Unit,
     input: @Composable RowScope.(MessageComposerState) -> Unit,
 ) {
-    val (value, attachments, activeAction, validationErrors) = messageComposerState
+    val (value, attachments, activeAction, validationErrors, mentionSuggestions, cooldownTimer) = messageComposerState
 
-    showValidationErrorIfNecessary(validationErrors)
+    MessageInputValidationError(validationErrors)
 
     Surface(
         modifier = modifier,
@@ -165,26 +197,78 @@ public fun MessageComposer(
 
                 input(messageComposerState)
 
-                val isInputValid = (value.isNotEmpty() || attachments.isNotEmpty()) && validationErrors.isEmpty()
+                if (cooldownTimer > 0) {
+                    CooldownIndicator(cooldownTimer = cooldownTimer)
+                } else {
+                    val isInputValid = (value.isNotEmpty() || attachments.isNotEmpty()) && validationErrors.isEmpty()
 
-                IconButton(
-                    enabled = isInputValid,
-                    content = {
-                        Icon(
-                            painter = painterResource(id = R.drawable.stream_compose_ic_send),
-                            contentDescription = stringResource(id = R.string.stream_compose_send_message),
-                            tint = if (isInputValid) ChatTheme.colors.primaryAccent else ChatTheme.colors.textLowEmphasis
-                        )
-                    },
-                    onClick = {
-                        if (isInputValid) {
-                            onSendMessage(value, attachments)
+                    IconButton(
+                        enabled = isInputValid,
+                        content = {
+                            Icon(
+                                painter = painterResource(id = R.drawable.stream_compose_ic_send),
+                                contentDescription = stringResource(id = R.string.stream_compose_send_message),
+                                tint = if (isInputValid) ChatTheme.colors.primaryAccent else ChatTheme.colors.textLowEmphasis
+                            )
+                        },
+                        onClick = {
+                            if (isInputValid) {
+                                onSendMessage(value, attachments)
+                            }
                         }
-                    }
-                )
+                    )
+                }
             }
         }
+
+        if (mentionSuggestions.isNotEmpty()) {
+            mentionPopupContent(mentionSuggestions)
+        }
     }
+}
+
+/**
+ * Represent a timer that show the remaining time until the user is allowed to send the next message.
+ *
+ * @param cooldownTimer The amount of time left until the user is allowed to sent the next message.
+ * @param modifier Modifier for styling.
+ */
+@Composable
+internal fun CooldownIndicator(
+    cooldownTimer: Int,
+    modifier: Modifier = Modifier,
+) {
+    Box(
+        modifier = modifier
+            .size(48.dp)
+            .padding(12.dp)
+            .background(shape = RoundedCornerShape(24.dp), color = ChatTheme.colors.disabled),
+        contentAlignment = Alignment.Center
+    ) {
+        Text(
+            text = cooldownTimer.toString(),
+            color = Color.White,
+            textAlign = TextAlign.Center,
+            style = ChatTheme.typography.bodyBold
+        )
+    }
+}
+
+/**
+ * Represents the default mention suggestion list popup shown above the message composer.
+ *
+ * @param mentionSuggestions The list of users that can be used to autocomplete the current mention input.
+ * @param onMentionClick Handler when the user taps on a mention suggestion item.
+ */
+@Composable
+internal fun DefaultMentionPopupContent(
+    mentionSuggestions: List<User>,
+    onMentionClick: (User) -> Unit,
+) {
+    MentionSuggestionList(
+        users = mentionSuggestions,
+        onMentionClick = { onMentionClick(it) }
+    )
 }
 
 /**
@@ -208,7 +292,7 @@ internal fun DefaultComposerLabel() {
  * @param validationErrors The list of validation errors for the current user input.
  */
 @Composable
-private fun showValidationErrorIfNecessary(validationErrors: List<ValidationError>) {
+private fun MessageInputValidationError(validationErrors: List<ValidationError>) {
     if (validationErrors.isNotEmpty()) {
         val errorMessage = when (val validationError = validationErrors.first()) {
             is ValidationError.MessageLengthExceeded -> {