+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/theme/StreamDimens.kt
@@ -23,6 +23,12 @@ import androidx.compose.ui.unit.dp
  * @param threadSeparatorVerticalPadding The vertical content padding inside thread separator item.
  * @param threadSeparatorTextVerticalPadding The vertical padding inside thread separator text.
  * @param messageOverlayActionItemHeight The height of an action item on the selected message overlay.
+ * @param suggestionListMaxHeight The maximum height of the suggestion list popup.
+ * @param suggestionListPadding The outer padding of the suggestion list popup.
+ * @param suggestionListElevation THe elevation of the suggestion list popup.
+ * @param mentionSuggestionItemHorizontalPadding The horizontal content padding inside mention list item.
+ * @param mentionSuggestionItemVerticalPadding The vertical content padding inside mention list item.
+ * @param mentionSuggestionItemAvatarSize The size of a channel avatar in the suggestion list popup.
  * @param messageOptionsMaxHeight The max height of the message options section when we select a message in the list.
  * @param messageOptionsMaxWidth The max width of the message options section when we select a message in the list.
  * @param messageOptionsRoundedCorners The rounded corners size of the message options shape.
@@ -45,6 +51,12 @@ public data class StreamDimens(
     public val threadSeparatorVerticalPadding: Dp,
     public val threadSeparatorTextVerticalPadding: Dp,
     public val messageOverlayActionItemHeight: Dp,
+    public val suggestionListMaxHeight: Dp,
+    public val suggestionListPadding: Dp,
+    public val suggestionListElevation: Dp,
+    public val mentionSuggestionItemHorizontalPadding: Dp,
+    public val mentionSuggestionItemVerticalPadding: Dp,
+    public val mentionSuggestionItemAvatarSize: Dp,
     public val messageOptionsMaxWidth: Dp,
     public val messageOptionsMaxHeight: Dp,
     public val messageOptionsRoundedCorners: Dp
@@ -67,6 +79,12 @@ public data class StreamDimens(
             threadSeparatorVerticalPadding = 8.dp,
             threadSeparatorTextVerticalPadding = 2.dp,
             messageOverlayActionItemHeight = 40.dp,
+            suggestionListMaxHeight = 256.dp,
+            suggestionListPadding = 8.dp,
+            suggestionListElevation = 4.dp,
+            mentionSuggestionItemHorizontalPadding = 16.dp,
+            mentionSuggestionItemVerticalPadding = 8.dp,
+            mentionSuggestionItemAvatarSize = 40.dp,
             messageOptionsMaxWidth = 200.dp,
             messageOptionsMaxHeight = 300.dp,
             messageOptionsRoundedCorners = 16.dp