+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/common/markdown/ChatMarkdownImpl.kt
@@ -20,7 +20,7 @@ internal class ChatMarkdownImpl(context: Context) : ChatMarkdown {
         .build()
 
     override fun setText(textView: TextView, text: String) {
-        markwon.setMarkdown(textView, text)
+        markwon.setMarkdown(textView, fixItalicAtEnd(text))
         Linkify.addLinks(textView)
     }
 }